+++ custom_chunck/src/compute/app/document.py
@@ -1,5 +1,6 @@
 from shared_oci import log
 from shared_oci import log_in_file
+import shared
 import shared_oci
 import pathlib
 import os
@@ -29,33 +30,33 @@ def eventDocument(value):
     result = { "content": "-" }
     if resourceExtension in [".tif"] or resourceName.endswith(".anonym.pdf"):
         # This will create a JSON file in Object Storage that will create a second even with resourceExtension "json" 
-        shared_oci.documentUnderstanding(value)
+        shared_oci.convertOciDocumentUnderstanding(value)
         return
     elif resourceExtension in [".pdf", ".txt", ".csv", ".md", ""] or resourceName in ["_metadata_schema.json", "_all.metadata.json"] :
         # Simply copy the file to the agent bucket
-        shared_oci.upload_agent_bucket(value)
+        shared_oci.convertUpload(value)
         return
     # elif resourceExtension in [".png", ".jpg", ".jpeg", ".gif"]:
-    #    shared_oci.image2pdf(value)
+    #    shared_oci.convertImage2Pdf(value)
     #    return    
     elif resourceExtension in [".docx", ".doc",".pptx", ".ppt", ".html"]:
         shared_oci.docling2md(value)
         return        
     elif resourceExtension in [".mp3", ".mp4", ".avi", ".wav", ".m4a"]:
         # This will create a SRT file in Object Storage that will create a second even with resourceExtension ".srt" 
-        shared_oci.speech(value)
+        shared_oci.convertOciSpeech(value)
         return
     elif resourceExtension in [".sitemap"]:
         # This will create a PDFs file in Object Storage with the content of each site (line) ".sitemap" 
-        shared_oci.sitemap(value)
+        shared_oci.convertSitemapText(value)
         return   
     elif resourceExtension in [".crawler"]:
         # This will crawl all HTML pages of a website 
-        shared_oci.crawler(value)
+        shared_oci.convertCrawler(value)
         return    
     elif resourceExtension in [".webp"]:
         # Convert webp to PNG
-        shared_oci.webp2png(value)
+        shared_oci.convertWebp2Png(value)
         return
     elif resourceExtension in [".srt"]:
         log("IGNORE .srt")
@@ -67,21 +68,21 @@ def eventDocument(value):
 
     if eventType in [ "com.oraclecloud.objectstorage.createobject", "com.oraclecloud.objectstorage.updateobject" ]:
         if resourceExtension in [".json"]:
-            result = shared_oci.decodeJson(value)
+            result = shared_oci.convertJson(value)
         elif resourceExtension in [".png", ".jpg", ".jpeg", ".gif"]:
-            result = shared_oci.vision(value)
+            result = shared_oci.convertOciVision(value)
         else:
-            result = shared_oci.invokeTika(value)
+            result = shared_oci.convertOciFunctionTika(value)
 
         if result:
             log_in_file("content", result["content"])
             if len(result["content"])==0:
                 return 
-            shared_oci.upload_agent_bucket(value, result["content"], result["path"])    
+            shared_oci.convertUpload(value, result["content"], result["path"])    
 
     elif eventType == "com.oraclecloud.objectstorage.deleteobject":
         # No need to get the content for deleting
-        shared_oci.upload_agent_bucket(value, "-")    
+        shared_oci.convertUpload(value, "-")    
 
 
     # Upload to the Agent Bucket
@@ -92,12 +93,13 @@ def eventDocument(value):
 countUpdate = 0
 
 def updateCount(count):
+    ## XXX Not needed for DB23ai ? And/or bulk calculate the Embedding via a PLSQL procedure ?
     global countUpdate
     if count>0:
         countUpdate = countUpdate + count 
     elif countUpdate>0:
         try:
-            shared_oci.genai_agent_datasource_ingest()
+            shared.genai_agent_datasource_ingest()
             log( "<updateCount>GenAI agent datasource ingest job created")
             countUpdate = 0
         except (Exception) as e:
